requirements.txt
torch>=2.0.0
torch-geometric==2.0.0
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
pytorch-lightning>=2.2.4
pytz==2024.2
PyYAML==6.0.1
numpy>=1.23.0
matplotlib>=3.7.0
tqdm>=4.65.0
h5py>=3.8.0
pytorch_msssim>=0.2.1
lpips>=0.1.4
opencv-python>=4.7.0
einops>=0.6.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.14.1
packaging==24.0
pandas==2.2.3
yacs==0.1.8
yapf==0.43.0
yarl==1.9.4
zipp==3.21.0
zlib-state==0.1.9